requirements.txt
aiogram==2.14.3
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
babel==2.17.0
billiard==4.2.0
blinker==1.8.2
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-simple-history==3.7.0
djangorestframework==3.15.2
docx==0.2.4
dotenv-python==0.0.1
environs==14.1.1
exceptiongroup==1.2.2
Flask==3.0.3
flower==2.0.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanize==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.26.1
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
